:construction: Moved schema gen
